Small changes just for fun
Make the background brighter.

When ball enters alert mode due to repeated straight bounce, so
as to switch to next level, the color of ball is changed as a
visual cue wrt same, so one can either decide to continue in
straight bounce to advance to next level, so else move paddle
a bit to remain in same level with a angular bounce.

Also if user avoids switching to next level, the repeated straight
bounce level switching tracker variable is reset.

THe rate at which the speed increases between levels has been
slowed down a bit.