Basic skeleton for mouse/touch input, larger canvas
THis is a basic skeleton for using mouse/touch. However this in
itself doesnt make it playable, which needs better translation
of mouse clicks to paddle movements (potentially dynamic variable
rate or so). However for now the canvas has been made larger along
y axis, to give more rope, to allow one to get along with atleast
the initial 1 odd levels.

For now keyboard is the best input device for this game.