Finer grained movement of paddle
Allows one to move paddle around easily either towards the
ball or away from it, with sensible speeds.

If near to ball vertical, then move with finer control (finer than
before with this commit), if away from ball then move in a coarse
grained manner.

This gives better braking effect as one nears the ball. And as always
there is the acceleration as one moves away from it.