Fix ArcOut wrt requestAnimationFrame flow
As requestAnimationFrame doesnt allow any additional arguments to
be passed to the call back, so use bound function to pass additional
argument. Using this the need to pass the callback to be called by
arc_out, once its done is taken care of now.

NOTE: The idea is arc_out can be generic and called either at the
end like in the current case, or if required at any other time
also, by passing a different callback to be called at the end.

Also as requestAnimFrame callbacks are called at a rate of around
60 fps or so in many cases, while I need arc_out to work with
around 10 fps, so also have reduced the rate of growth of the arc.
This is a hardcoded way of doing it rather than the ideal anim_adjust
mechanism which I had created, which also means that on systems
where the browser calls the callback at a slower rate, the arc_out
will occur slower.