Dont forget to request fresh anim frame, where required
anim_adjust returns 0 when there are two frames called with same
timestamp, which can be safely ignored without needing to request
a new anim frame, because the previous call with diff in timestamp
would have already requested the next anim frame.

However for special situations like a large gap between successive
frames, we want to ignore the animation adjustment logic, but we
still need to request the next anim frame, so handle that case now.