Pass canvas context around, wrt background scene logic
The logics which mainly generate the background of the scene, now
take a canvas context, so that they can either write to screen
or to a offscreen canvas.

If writing to offscreen canvas, then that canvas will be duplicated
into main screen before the actors in the scene ie paddle and ball
are drawn.