Exp: Use offscreen canvas, draw background and zone to it
Added a offscreen canvas, into which the background fill and
zone is now drawn. Inturn this is done only once for each
level, as it is entered. The offscreen canvas is inturn copied
into the main screen canvas.

Also avoid the translucent background fill, for now. The canvas
copy (drawImage, doesnt simulate equivalent for now).

Need to set flag to clear zone once t2l is done. Also need to
reduce t2l as required.